# Add performance benchmarks