--- a/.travis.yml
+++ b/.travis.yml
@@ -1,7 +1,7 @@
-sudo: false
 language: node_js
 node_js:
   - '10.16.0'
+os: 'linux'
 script:
   - pnpm run build && pnpm run test:ci
 after_script: