--- a/package.json
+++ b/package.json
@@ -17,15 +17,21 @@
   "license": "MIT",
   "devDependencies": {
     "@types/assert": "^1.4.4",
+    "@types/benchmark": "^1.0.31",
     "@types/mocha": "^5.2.7",
     "assert": "^2.0.0",
+    "benchmark": "github:bestiejs/benchmark.js#42f3b732bac3640eddb3ae5f50e445f3141016fd",
+    "callbag-basics": "^3.2.0",
     "fast-check": "^1.21.0",
     "mocha": "^7.0.1",
+    "most": "^1.8.0",
     "nyc": "^15.0.0",
     "prettier": "^1.19.1",
     "rollup": "^1.29.1",
+    "rxjs": "^6.5.4",
     "ts-node": "^8.6.2",
-    "typescript": "^3.7.5"
+    "typescript": "^3.7.5",
+    "xstream": "^11.11.0"
   },
   "prettier": {
     "singleQuote": true