--- a/perf/index.ts
+++ b/perf/index.ts
@@ -1,0 +1,4 @@
+import { runSuite } from './helpers';
+import { fold } from './fold';
+
+[fold].forEach(runSuite);