--- a/perf/index.ts
+++ b/perf/index.ts
@@ -1,4 +1,8 @@
 import { runSuite } from './helpers';
 import { fold } from './fold';
+import { dataflow } from './dataflow';
 
-[fold].forEach(runSuite);
+[fold, dataflow].reduce(
+  (acc, curr) => acc.then(() => runSuite(curr)),
+  Promise.resolve()
+);