--- a/src/constants.ts
+++ b/src/constants.ts
@@ -1,1 +1,0 @@
-export const undef = Symbol('undef');