--- a/src/skip.ts
+++ b/src/skip.ts
@@ -1,5 +1,4 @@
 import { Operator } from './types';
-import { undef } from './constants';
 
 export function skip<T>(n: number): Operator<T, T> {
   return source => (_, sink) => {
@@ -13,11 +12,11 @@ export function skip<T>(n: number): Operator<T, T> {
 
 export function last<T>(): Operator<T, T> {
   return source => (_, sink) => {
-    let last: any = undef;
+    let last: any[] = [];
 
     source(0, (t, d) => {
-      if (t === 1) last = d;
-      else if (t === 2 && last !== undef) sink(1, last);
+      if (t === 1) last[0] = d;
+      else if (t === 2 && last.length > 0) sink(1, last[0]);
       if (t !== 1) sink(t, d);
     });
   };