For `scan`:
```
scan -> scan 1000000 integers
-----------------------------------------------------
@cycle/callbags    25.00 op/s ±  1.15%   (61 samples)
callbag-basics     26.37 op/s ±  1.14%   (65 samples)
xstream            13.99 op/s ±  1.16%   (67 samples)
rxjs               15.20 op/s ±  1.79%   (73 samples)
most               28.18 op/s ±  1.77%   (67 samples)
-----------------------------------------------------
```