@jvanbruegge 2020-02-05
The difference between `@cycle/callbags` and `callbag-basics` does not really make sense because it is exactly the same code