@jvanbruegge 2020-02-06
I'd expect to see a difference with other operators. I also found a low-hanging performance fruit (see map.ts), so this is definetly not wasted